+++ src/main/java/org/github/springframework/beans/factory/DisposableBean.java
@@ -1,0 +1,9 @@
+package org.github.springframework.beans.factory;
+
+/**
+ * @author zhaoyu
+ * Created on 2023-11-27
+ */
+public interface DisposableBean {
+    void destroy() throws Exception;
+}