+++ src/main/java/org/github/springframework/beans/factory/InitializingBean.java
@@ -1,0 +1,10 @@
+package org.github.springframework.beans.factory;
+
+/**
+ * @author zhaoyu
+ * Created on 2023-11-27
+ */
+public interface InitializingBean {
+    void afterPropertiesSet() throws Exception;
+
+}