+++ src/main/java/org/github/springframework/beans/factory/config/ConfigurableBeanFactory.java
@@ -12,4 +12,10 @@ public interface ConfigurableBeanFactory  extends HierarchicalBeanFactory, Singl
      * @param beanPostProcessor
      */
     void addBeanPostProcessor(BeanPostProcessor beanPostProcessor);
+
+
+    /**
+     * 销毁单例bean
+     */
+    void destroySingletons();
 }