+++ src/main/java/org/github/springframework/beans/factory/support/DefaultSingletonBeanRegistry.java
@@ -2,7 +2,10 @@ package org.github.springframework.beans.factory.support;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Set;
 
+import org.github.springframework.beans.BeansException;
+import org.github.springframework.beans.factory.DisposableBean;
 import org.github.springframework.beans.factory.config.SingletonBeanRegistry;
 
 /**
@@ -12,6 +15,9 @@ import org.github.springframework.beans.factory.config.SingletonBeanRegistry;
 public class DefaultSingletonBeanRegistry implements SingletonBeanRegistry {
     private Map<String, Object> singletonObjects = new HashMap<>();
 
+    private final Map<String, DisposableBean> disposableBeans = new HashMap<>();
+
+
     @Override
     public Object getSingleton(String beanName) {
         return singletonObjects.get(beanName);
@@ -19,5 +25,21 @@ public class DefaultSingletonBeanRegistry implements SingletonBeanRegistry {
 
     protected void addSingleton(String beanName, Object singletonObject) {
         singletonObjects.put(beanName, singletonObject);
+    }
+
+    public void registerDisposableBean(String beanName, DisposableBean bean) {
+        disposableBeans.put(beanName, bean);
+    }
+
+    public void destroySingletons() {
+        Set<String> beanNames = disposableBeans.keySet();
+        for (String beanName : beanNames) {
+            DisposableBean disposableBean = disposableBeans.remove(beanName);
+            try {
+                disposableBean.destroy();
+            } catch (Exception e) {
+                throw new BeansException("Destroy method on bean with name '" + beanName + "' threw an exception", e);
+            }
+        }
     }
 }