+++ changelog.md
@@ -78,3 +78,22 @@ BeanFactory是spring的基础设施，面向spring本身；而ApplicationContext
 从bean的角度看，目前生命周期如下：
 
 ![](./assets/application-context-life-cycle.png)
+
+
+## bean的初始化和销毁方法
+> 分支：init-and-destroy-method
+
+在spring中，定义bean的初始化和销毁方法有三种方法：
+- 在xml文件中制定init-method和destroy-method
+- 继承自InitializingBean和DisposableBean
+- 在方法上加注解PostConstruct和PreDestroy
+
+第三种通过BeanPostProcessor实现，在扩展篇中实现，本节只实现前两种。
+
+针对第一种在xml文件中指定初始化和销毁方法的方式，在BeanDefinition中增加属性initMethodName和destroyMethodName。
+
+初始化方法在AbstractAutowireCapableBeanFactory#invokeInitMethods执行。DefaultSingletonBeanRegistry中增加属性disposableBeans保存拥有销毁方法的bean，拥有销毁方法的bean在AbstractAutowireCapableBeanFactory#registerDisposableBeanIfNecessary中注册到disposableBeans中。为了确保销毁方法在虚拟机关闭之前执行，向虚拟机中注册一个钩子方法，查看AbstractApplicationContext#registerShutdownHook（非web应用需要手动调用该方法）。当然也可以手动调用ApplicationContext#close方法关闭容器。
+
+到此为止，bean的生命周期如下：
+
+![](./assets/init-and-destroy-method.png)